+++ src/main/php/filter/MonthFilter.php
@@ -42,9 +42,9 @@ class MonthFilter implements Filter
         }
 
         try {
-            return Month::fromString($param->getValue());
+            return Month::fromString($param->value());
         } catch (IllegalArgumentException $iae) {
-            $param->addErrorWithId('MONTH_INVALID');
+            $param->addError('MONTH_INVALID');
         }
 
         return null;